Number of Steps to Reduce a Number in Binary Representation to One

+++ number-of-steps-to-reduce-a-number-in-binary-representation-to-one/Solution.java
@@ -10,7 +10,7 @@ class Solution {
     }
 
     if (index != 0) {
-      result += index + (int) IntStream.range(0, index).filter(i -> s.charAt(i) == '0').count() + 2;
+      result += index + IntStream.range(0, index).filter(i -> s.charAt(i) == '0').count() + 2;
     }
 
     return result;
